completed reply functionality

+++ controllers/feed/dataController.js
@@ -1,10 +1,11 @@
 const Post = require("../../models/post")
+const Reply = require("../../models/reply");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 
 exports.index = async (req, res, next) => {
     try {
-        res.locals.data.posts = await Post.find({}).populate("author", "shortId");
+        res.locals.data.posts = await Post.find({}).populate("author", "shortId").sort({ createdAt: -1 })
         next();
     }
     catch (error) {
@@ -14,6 +15,12 @@ exports.index = async (req, res, next) => {
 
 exports.deletePost = async (req, res, next) => {
     try {
+        const post = await Post.findById(req.params.id);
+
+        if (post.author.toString() !== req.user._id.toString()) {
+            return res.status(403).send({ message: "Not authorized to delete this post"})
+        }
+
         await Post.findOneAndDelete({ "_id": req.params.id })
         .then(() => {
             next();
@@ -26,6 +33,12 @@ exports.deletePost = async (req, res, next) => {
 
 exports.updatePost = async (req, res, next) => {
     try {
+         const post = await Post.findById(req.params.id);
+
+        if (!req.user || post.author.toString() !== req.user._id.toString()) {
+            return res.status(403).send({ message: "Not authorized to edit this post"})
+        }
+
         const updatedPost = await Post.findByIdAndUpdate(req.params.id,  
             {
                 type: req.body.type,
@@ -33,7 +46,7 @@ exports.updatePost = async (req, res, next) => {
                 image: req.body.image,
                 text: req.body.text,
             },
-            { new: true }).populate("author", "color");
+            { new: true }).populate("author", "shortId")
             res.locals.data.post = updatedPost;
         next();
     }
@@ -62,10 +75,21 @@ exports.createPost = async (req, res, next) => {
 
 exports.showPost = async (req, res, next) => {
     try {
-        res.locals.data.post = await Post.findById(req.params.id).populate("author", "shortId")
+        
+        res.locals.data.post = await Post.findById(req.params.id).populate("author", "shortId _id")
         if (!res.locals.data.post) {
             throw new Error(`No post with id ${req.params.id} in our database`);
         }
+
+        if(req.path.includes("edit")) {
+            if (!req.user || post.author.toString() !== req.user._id.toString()) {
+            return res.status(403).send({ message: "Not authorized to edit this post"})
+        }
+        }
+
+        const replies = await Reply.find({ post: req.params.id }).populate("author", "shortId").sort({ createdAt: -1 })
+        console.log("Replies fetched:", replies.length);
+        res.locals.data.replies = replies;
         next();
     }
     catch (error) {
@@ -75,7 +99,7 @@ exports.showPost = async (req, res, next) => {
 
 exports.asks = async (req, res, next) => {
     try {
-        res.locals.data.posts = await Post.find({ type: "ask" }).populate("author", "shortId")
+        res.locals.data.posts = await Post.find({ type: "ask" }).populate("author", "shortId").sort({ createdAt: -1 })
         next();
     }
     catch (error) {
@@ -85,7 +109,7 @@ exports.asks = async (req, res, next) => {
 
 exports.feels = async (req, res, next) => {
     try {
-        res.locals.data.posts = await Post.find({ type: "feel" }).populate("author", "shortId")
+        res.locals.data.posts = await Post.find({ type: "feel" }).populate("author", "shortId").sort({ createdAt: -1 })
         next();
     }
     catch (error) {
